api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg[binary]==3.1.16
pgvector==0.2.4
pydantic==2.5.3
python-multipart==0.0.6
boto3==1.34.18
redis==5.0.1
rq==1.15.1
python-docx==1.1.0
pdfminer.six==20221105
pytest==7.4.4
pytest-asyncio==0.23.3
python-dotenv==1.0.0
cryptography==41.0.7
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
httpx==0.26.0
alembic==1.13.1
numpy==2.1.2
pyyaml==6.0.1
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-otlp==1.22.0
email-validator==2.1.0.post1
